requirements.txt
streamlit
gspread
oauth2client
twilio
python-dotenv
pandas
